flask_demo/requirements.txt
Flask==2.2.3
matplotlib==3.5.3
numpy==1.24.2
opencv_python==4.7.0.72
pandas==1.4.4
Pillow==9.4.0
scikit_learn==1.2.1
seaborn==0.12.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.1
